+++ src/utils/db.js
@@ -19,6 +19,11 @@ const news = mongoose.model('new', mongoose.Schema({
   title: String,
   desc: String,
   path: String,
+  pullRequest: [{
+    _id: mongoose.Types.ObjectId,
+    personId: String,
+    diff: String,
+  }],
   author: mongoose.Types.ObjectId,
   date: Date,
   dateLastmod: Date,
@@ -31,6 +36,19 @@ const news = mongoose.model('new', mongoose.Schema({
     isBasedOn: [String],
   },
   content: String,
+  comments: [{
+    _id: mongoose.Types.ObjectId,
+    personId: String,
+    content: String,
+    comments: [{
+      _id: mongoose.Types.ObjectId,
+      personId: String,
+      content: String,
+      reactions: [[String, Number]],
+    }],
+    reactions: [[String, Number]],
+  }],
+  reactions: [[String, Number]],
 }, { versionKey: false }));
 
 const userSchema = mongoose.Schema({
@@ -46,12 +64,14 @@ const userSchema = mongoose.Schema({
     gender: Number,
   },
   password: String,
+  roles: [String],
 }, { versionKey: false });
 
 userSchema.post('findOne', (person) => {
   if (person !== null) delete person._doc.password;
 });
 
+// eslint-disable-next-line func-names
 userSchema.pre('save', function (next) {
   if (this.password && this.isModified('password')) {
     this.password = createHash('sha256').update(this.password).digest('hex');